[DEMO-2561] Remove setup_database_method, replace with class
* Created new class SetupDatabaseAction, to be a single place for the
functionality of setting up the database. This lets me remove the 2x
methods I previously had (one for lambda, one for tron). Reduces code
duplication.
