+++ python/lambda.py
@@ -1,9 +1,6 @@
 import json
 import os
 
-import mysql.connector
-from mysql.connector import errorcode
-
 from api.help import help_message
 from api.index_handler import IndexHandler
 from api.inventory_handler import InventoryHandler
@@ -18,76 +15,8 @@ from repository.database_connection_info import DatabaseConnectionInfo
 from repository.database_inventory_repository import \
     DatabaseInventoryRepository
 from repository.file_inventory_repository import FileInventoryRepository
-from repository.i_database_connector import IDatabaseConnector
 from repository.mysql_connector import MySqlConnector
-
-
-def setup_database(database_connector: IDatabaseConnector, database_connection_info: DatabaseConnectionInfo, inventory_data):
-    DB_NAME = database_connection_info.database
-    # remove this since we dont know if database exists yet.
-    database_connection_info.database = None
-
-    cnx = database_connector.connect(**database_connection_info.asdict())
-    cursor = cnx.cursor()
-
-    def create_database(cursor):
-        try:
-            cursor.execute(
-                "CREATE DATABASE {} DEFAULT CHARACTER SET 'utf8'".format(DB_NAME))
-        except mysql.connector.Error as err:
-            print("Failed creating database: {}".format(err))
-            exit(1)
-
-    try:
-        cursor.execute("USE {}".format(DB_NAME))
-        print("Database already exists")
-    except mysql.connector.Error as err:
-        print("Database {} does not exists.".format(DB_NAME))
-        if err.errno == errorcode.ER_BAD_DB_ERROR:
-            create_database(cursor)
-            print("Database {} created successfully.".format(DB_NAME))
-            cnx.database = DB_NAME
-
-            TABLES = {}
-            TABLES['inventory'] = (
-                "CREATE TABLE `inventory` ("
-                "  `id` int(11) NOT NULL AUTO_INCREMENT,"
-                "  `item` varchar(14) NOT NULL,"
-                "  `sku` varchar(80) NOT NULL,"
-                "  `price` varchar(14) NOT NULL,"
-                "  PRIMARY KEY (`id`)"
-                ") ENGINE=InnoDB")
-
-            print("Creating inventory table . . .")
-
-            for table_name in TABLES:
-                table_description = TABLES[table_name]
-                try:
-                    print("Creating table {}: ".format(table_name), end='')
-                    cursor.execute(table_description)
-                except mysql.connector.Error as err:
-                    if err.errno == errorcode.ER_TABLE_EXISTS_ERROR:
-                        print("already exists.")
-                    else:
-                        print(err.msg)
-                else:
-                    print("OK")
-
-            print("Inserting inventory data . . .")
-
-            add_inventory_item = ("INSERT INTO inventory "
-                                  "(item, price, sku) "
-                                  "VALUES (%(item)s, %(price)s, %(sku)s)")
-
-            for item in inventory_data:
-                cursor.execute(add_inventory_item, item)
-                cnx.commit()
-
-            cursor.close()
-            cnx.close()
-        else:
-            print(err)
-            exit(1)
+from repository.setup_database_action import SetupDatabaseAction
 
 
 def get_config_file_path(arguments):
@@ -176,7 +105,8 @@ def lambda_handler(event, context):
     AppLogging.init('info')
     
     if(inventory_repository_selector(app_config) == 'database'):
-        setup_database(database_connector, database_connection_info_func(), get_inventory_data())
+        setup_database = SetupDatabaseAction(database_connector, database_connection_info_func(), get_inventory_data())
+        setup_database.execute()
 
     endpoint = (event['resource'] or "").lower()
     if endpoint == "/":
